refactor(86byb8er1): moved requests and responses to objects package
