+++ src/main/java/pl/wip/wipplatformbe/objects/requests/RegisterRequest.java
@@ -1,4 +1,4 @@
-package pl.wip.wipplatformbe.requests;
+package pl.wip.wipplatformbe.objects.requests;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
