+++ src/main/java/pl/wip/wipplatformbe/objects/requests/TestRequest.java
@@ -1,0 +1,5 @@
+package pl.wip.wipplatformbe.objects.requests;
+
+public record TestRequest(String name) {
+
+}
