+++ src/main/java/pl/wip/wipplatformbe/objects/responses/AuthenticationResponse.java
@@ -1,4 +1,4 @@
-package pl.wip.wipplatformbe.responses;
+package pl.wip.wipplatformbe.objects.responses;
 
 public record AuthenticationResponse(String username, String email, String token) {
 }
