--- src/main/java/pl/wip/wipplatformbe/requests/TestRequest.java
@@ -1,5 +1,0 @@
-package pl.wip.wipplatformbe.requests;
-
-public record TestRequest(String name) {
-
-}
